chore: moves fns into transform sub-mod
