+++ crates/css_ast/src/types/transform/mod.rs
@@ -1,19 +1,45 @@
+mod matrix;
+mod rotate;
+mod scale;
+mod scale_x;
+mod scale_y;
+mod skew;
+mod skew_x;
+mod skew_y;
+mod translate;
+
+use skew::SkewKind;
+
 use css_lexer::Cursor;
 use css_parse::{Parser, Peek, function_set};
-use crate::types;
-use csskit_derives::{Parse, ToCursors};
+use csskit_derives::{ToCursors, Parse};
 
 function_set!(TransformFunctionName {
 	Matrix: "matrix",
+	// Translate: "translate",
+	// TranslateX: "translateX",
+	// TranslateY: "translateY",
+	Scale: "scale",
+	ScaleX: "scaleX",
+	ScaleY: "scaleY",
 	Rotate: "rotate",
+	Skew: "skew",
+	SkewX: "skewX",
+	SkewY: "skewY",
 });
 
 // https://drafts.csswg.org/css-transforms-1/#two-d-transform-functions
 #[derive(Parse, ToCursors, Debug, Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
 #[cfg_attr(feature = "serde", derive(serde::Serialize), serde())]
 pub enum TransformFunction {
-	Matrix(types::Matrix),
-	Rotate(types::Rotate),
+	Matrix(matrix::Matrix),
+	Scale(scale::Scale),
+	ScaleX(scale_x::ScaleX),
+	ScaleY(scale_y::ScaleY),
+	Rotate(rotate::Rotate),
+	Skew(skew::Skew),
+	SkewX(skew_x::SkewX),
+	SkewY(skew_y::SkewY),
 }
 
 impl<'a> Peek<'a> for TransformFunction {
@@ -34,7 +60,14 @@ mod tests {
 
 	#[test]
 	fn test_writes() {
-		assert_parse!(TransformFunction, "rotate(45deg)");
 		assert_parse!(TransformFunction, "matrix(1,0,0,1,0,0)");
+		assert_parse!(TransformFunction, "scale(2)");
+		assert_parse!(TransformFunction, "scale(1,2)");
+		assert_parse!(TransformFunction, "scaleX(2)");
+		assert_parse!(TransformFunction, "scaleY(2)");
+		assert_parse!(TransformFunction, "rotate(45deg)");
+		assert_parse!(TransformFunction, "skew(1deg,2deg)");
+		assert_parse!(TransformFunction, "skewX(1deg)");
+		assert_parse!(TransformFunction, "skewY(1deg)");
 	}
 }
